Converted the data type of AffectedRows to long.

+++ OracleSqlBuilder/OracleSqlNonData.cs
@@ -9,7 +9,7 @@ namespace OracleSqlBuilder {
 		/// <summary>
 		/// The number of affected rows.
 		/// </summary>
-		public double AffectedRows {
+		public long AffectedRows {
 			get;
 			internal set;
 		}
